+++ extras/tests/JsonVariant/containsKey.cpp
@@ -26,3 +26,19 @@ TEST_CASE("JsonVariant::containsKey()") {
     REQUIRE(var.containsKey(std::string("world")) == false);
   }
 }
+
+TEST_CASE("JsonVariantConst::containsKey()") {
+  DynamicJsonDocument doc(4096);
+  doc["hello"] = "world";
+  JsonVariantConst cvar = doc.as<JsonVariant>();
+
+  SECTION("containsKey(const char*) returns true") {
+    REQUIRE(cvar.containsKey("hello") == true);
+    REQUIRE(cvar.containsKey("world") == false);
+  }
+
+  SECTION("containsKey(std::string) returns true") {
+    REQUIRE(cvar.containsKey(std::string("hello")) == true);
+    REQUIRE(cvar.containsKey(std::string("world")) == false);
+  }
+}