+++ CHANGELOG.md
@@ -5,6 +5,8 @@ HEAD
 ----
 
 * Fixed regression in UTF16 decoding (issue #1173)
+* Fixed `containsKey()` on `JsonVariantConst`
+* Added `getElement()` and `getMember()` to `JsonVariantConst`
 
 v6.14.0 (2020-01-16)
 -------